Changed the book class to represent a complete book
Converted the book class to represented  complete book rather than the incomplete book information inside the book table on the database. It was decided that using this version of the book that was created during the query of the database was more effective and would meet the needs of the website barring additional details which may be added in the future.