Added queries for new API calls
Added the needed methods for the controller to be able to query the database for the information that  is needed by the website.
Also created a "baseQuery" for use in any other query that will need it. This can be added on to in order to create more specific calls if needed.