Converted all queries to stored procedures.
Given the unique circumstances of the "FullBook" entity. calling custom methods is a requirement. These have now been moved entirely to the database so that the repository class only calls the a stored procedure instead of having large amounts of query code located within the backend application. This should allow for much cleaner code and a theoretical improvement in security.